requirements.txt
fastapi 
uvicorn[standard] 
python-multipart 
openai==1.43.0 
httpx==0.27.0 
python-dotenv 
pdfplumber 
pytesseract 
pillow 
pydantic
pymupdf
pdfplumber
pdf2image
pandas
openpyxl
sqlalchemy
python-jose
passlib[bcrypt]
python-jose[cryptography]
passlib[bcrypt]
python-multipart